+++ include/tensor.h
@@ -68,9 +68,9 @@ void kernel_into_im2col_fwise(Tensor4D* kernel, bool flipped, Matrix* kernel_im2
 void kernel_into_im2col_chwise(Tensor4D* kernel, bool flipped, Matrix* kernel_im2col);
 void delta_into_im2col_fwise(Tensor4D* delta, int filter_idx, Matrix* im2col);
 void input_into_im2col_fwise(Tensor4D* input, int filter_idx, Tensor4D* kernel, int stride, int padding, Matrix* im2col);
-unsigned long tensor3D_get_sizeof_mem_allocated(Tensor3D* t);
-unsigned long tensor4D_get_sizeof_mem_allocated(Tensor4D* t);
-unsigned long tensor4D_uint16_get_sizeof_mem_allocated(Tensor4D_uint16* t);
+size_t tensor3D_get_sizeof_mem_allocated(Tensor3D* t);
+size_t tensor4D_get_sizeof_mem_allocated(Tensor4D* t);
+size_t tensor4D_uint16_get_sizeof_mem_allocated(Tensor4D_uint16* t);
 
 Tensor3D_uint16* tensor3D_uint16_new(int n_rows, int n_cols, int n_channels);
 void tensor3D_uint16_free(Tensor3D_uint16* t);
